More explicit casts

+++ bsa.c
@@ -9,13 +9,13 @@ underflowing. */
 
 unsigned addu32(unsigned const x, unsigned const y) {
 	unsigned res = x + y;
-	res |= (unsigned) (-(int) (res < x));
+	res |= (unsigned) (-(unsigned) (res < x));
 	return res;
 }
 
 unsigned subu32(unsigned const x, unsigned const y) {
 	unsigned res = x - y;
-	res &= (unsigned) (-(int) (res <= x));
+	res &= (unsigned) (-(unsigned) (res <= x));
 	return res;
 }
 
@@ -27,20 +27,20 @@ unsigned divu32(unsigned const x, unsigned const y) {
 unsigned mulu32(unsigned const x, unsigned const y) {
 	unsigned long const res = (unsigned long) x * (unsigned long) y;
 	unsigned const hi = (unsigned) (res >> 32);
-	return (unsigned) (res | (unsigned long) (-(int) (bool) hi));
+	return (unsigned) (res | (unsigned long) (-(unsigned) (bool) hi));
 }
 
 /* unsigned 64-bit */
 
 unsigned long addu64(unsigned long const x, unsigned long const y) {
 	unsigned long res = x + y;
-	res |= (unsigned long) (-(int) (res < x));
+	res |= (unsigned long) (-(unsigned long) (res < x));
 	return res;
 }
 
 unsigned long subu64(unsigned long const x, unsigned long const y) {
 	unsigned long res = x - y;
-	res &= (unsigned long) (-(res <= x));
+	res &= (unsigned long) (-(unsigned long) (res <= x));
 	return res;
 }
 
@@ -52,7 +52,7 @@ unsigned long divu64(unsigned long const x, unsigned long const y) {
 unsigned long mulu64(unsigned long const x, unsigned long const y) {
 	__uint128_t const res = (__uint128_t) x * (__uint128_t) y;
 	unsigned long const hi = (unsigned long) (res >> 64);
-	return (unsigned long) (res | (__uint128_t) (-(bool) hi));
+	return (unsigned long) (res | (__uint128_t) (-(unsigned long) (bool) hi));
 }
 
 /* signed 32-bit */
@@ -95,7 +95,7 @@ signed divs32(signed x, signed const y) {
 
 signed muls32(signed const x, signed const y) {
 	signed long res = (signed long) x * (signed long) y;
-	unsigned const res2 = ((unsigned) (x ^ y) >> 31) + INT_MAX;
+	unsigned const res2 = (((unsigned) x ^ (unsigned) y) >> 31) + INT_MAX;
 
 	signed const hi = (signed) (res >> 32);
 	signed const lo = (signed) res;
@@ -147,7 +147,8 @@ signed long divs64(signed long x, signed long const y) {
 
 signed long muls64(signed long const x, signed long const y) {
 	__int128_t res = (__int128_t) x * (__int128_t) y;
-	unsigned long const res2 = ((unsigned long) (x ^ y) >> 63) + LONG_MAX;
+	unsigned long const res2 = (((unsigned long) x ^ (unsigned long) y) >> 63)
+		+ LONG_MAX;
 
 	signed long const hi = (signed long) (res >> 64);
 	signed long const lo = (signed long) res;
